api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.27
pydantic>=2.6.1
psycopg2-binary>=2.9.9
alembic>=1.13.1
pytest>=8.0.0
httpx>=0.27.0
pgvector>=0.2.3
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9